MAINT: Refactor schedule header
